backend/requirements.txt
django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=4.0.0,<5.0.0
Pillow>=9.5.0,<10.0.0
